Revert "MicrodroidTest: Relax bccIsSuperficiallyWellFormed check"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/460394ef-2977-4fe5-801e-c15de34a332f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib9b29b8a73b85490044cd102c174da4005ac3b59
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com