Revert "Add microfuchsia apex that runs in AVF on bootup"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e1c0223f-4063-4c41-bafe-014da5351387). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 361108146
Change-Id: I3d048875df4ace7fc9a334cc0e633b12fbe57d18
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com