Revert "Append missing '",' in the config file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/db9b97b9-2a61-4463-bc23-f601a7a8ee2a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic99eb74236d77d21328bb417031655d5cd41166a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com