Framework JNI uses libvmclient to connect to virtmgr
libvmclient, as its name shows, is a library designed as a client-side
library for AVF. It is responsible for spawning virtmgr and make a
connection to it.

However, the library wasn't used for the Java APIs. For those APIs, the
JNI part (libvirtualization_jni) implemented the almost identical
routine for spawning virtmgr.

This change de-duplicates the code in the JNI part by exposing an FFI
for libvmclient and let the JNI part use it.

Note that the JNI part had more checks at the end of the connection
establishment. That part is copied over to libvmclient.

Bug: N/A
Test: watch TH

Change-Id: Ib6b0f3a180a7dbd4c797188e98d4b0e66f76423b