Revert "Add "stop vm" menu in Terminal"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d8753425-4a71-4dd7-a5c9-98e6e2307d1a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 362930190
Bug: 362931377
Change-Id: Ie244abf945016aa660bac1a5dbb9b329eeec8bbf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com