Revert "Don't wait vsock server if vm isn't running"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d8753425-4a71-4dd7-a5c9-98e6e2307d1a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 362930190
Bug: 362931377
Change-Id: I6f39bed78c2d1c0c46920933ebcb77bac7309ee7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com