Revert "Don't wait vsock server if vm isn't running"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fa45b302-bdf0-45f9-8ff3-0a0e578ea04b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 362964513
Bug: 362970373
Bug: 362971127
Bug: 362972750
Bug: 362972595
Bug: 362980192
Bug: 362986055
Bug: 362985780
Change-Id: I006db20f3fdf1fdb40ebe827c7596d5e49a6daba
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com