Revert "Add "stop vm" menu in Terminal"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5a984cc2-3bc4-4ec8-9c87-ed85095aa40c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie0d958055f4b257660cc043487cd4bab11eff888
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com