Revert "vm: DRY *Config structs w.r.t. aflags using cfg_if"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/236ee0f9-1047-404e-87c9-946c8bfdc566). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 364295749
Change-Id: Idcbc550da886d8757a97c4315749e2f080e2b8e7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com