Revert "Remove the use of from_raw_fd from fd_server/aidl.rs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d467a76e-cc1d-4a29-9c06-b1227b226ee4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I49a6cf2d5a9de2f2715953154b5c81c6c3e811f5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com