Revert "Add early_vms xsd schema"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/785ae138-d836-4597-b0c7-fd8fc64866cc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I184abc896a83b132929892e606eec5e64c35fb0a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com