Revert "Implement early_virtmgr"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/785ae138-d836-4597-b0c7-fd8fc64866cc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4fd8b5a884d5f965ef9651e916febd7a169fad07
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com