Revert "Remove the use of from_raw_fd from fd_server/aidl.rs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3dea295f-fe38-4ee9-a786-293216d2146b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2c428fd9563cb9e61791533c4f963491ff7f07f9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com