Revert "Update IAccessor implementation for getInstance"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0d0d786c-4ddb-4e59-8835-cc738733c1df). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I189ae5831d6c095cf15aacae673d9e883efc1905
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com