Revert "Take ownership of inherited FDs using the rustutils crate"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/82bfede8-544b-4895-b148-6d8d8de0ade8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 369012446
Change-Id: Iddd6c92e40770e158e82264b4b8656ca68c8ae76
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com