pvmfw: Detect and log guest UEFI support
Detect and log UEFI support for the guest kernel in main.rs. Duplicate
the Microdroid image - add new microdroid Soong target with EFI support.
The purpose of copying microdroid image is to avoid production
breakages.

Add tests to verify the EFI boot.

Test: ?
Built and installed  AVF and pvmfw to the device in order to perform EFI support flag existance check.

Change-Id: I60694425ebd280ca743b5a861d1241641c6071cb