Mark service-compos as system_ext_specific
Although service-compos is not installed in /system_ext, it is an apex
system server jar. These jars are preopted and the files are installed
outside the apex. All such files were previously installed in /system/,
but with https://r.android.com/3288083, these will be installed in the
same partition as the parent apex.

Test: WIP
Bug: TODO
Change-Id: Id0e7b462881bc24c94e60102d9c67b48a0787243