Revert "Link to loopXpY symbolically"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/df7816d4-fd2e-44f6-b09b-f5aa71e6c86f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibc511792a9cb38543382957eb649fc28613162d2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com