Revert "Link to loopXpY symbolically"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dafda2d9-d4a4-4594-9f81-7145d7449606). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I55dc5eef85e28221372a0fdbd690968485b1c39c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com