Revert "Link to loopXpY symbolically"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/87f4a8af-7556-4221-955f-83693e6e58d3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic10e2ed7d256609a1b4a44c1cd783a0ba2423105
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com