Revert "Link to loopXpY symbolically"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9fc07a95-aaa5-47cd-ba7c-2cd68a7bbf2b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1d141d9cf5b75fbfdb34a3264ad55894efc4d984
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com