Revert "Link to loopXpY symbolically"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8d0a4fa8-b91c-4ce0-b07c-d1039d08594a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5c5d5499e27592491d40565f23a8115fd81a58bc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com