Revert "Link to loopXpY symbolically"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/09f6235b-9e29-44fa-8a1f-5eebcf8372ec). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I29e8df7839f45ec0468f506fcbf6f3ba610de208
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com