Revert "Link to loopXpY symbolically"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c514c54d-aaa6-4e8a-ab0d-d35b6505b67e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I79b479eb4bc28296de3ab0726eef83e94e920038
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com