vmbase: uart: Use arch-agnostic write_volatile()
Replace the manual assembly with the arch-agnostic write_volatile(),
making the code more portable with the guarantee that rustc will
perform the same memory accesses, from [1]:

  The semantics will almost always end up pretty similar to C11’s
  definition of volatile. The compiler shouldn’t change the relative
  order or number of volatile memory operations.

Interestingly, because rustc is more aware of what write_volatile()
does, compared to the asm! block, it even seems to be able to further
optimize write_str() when unrolling the loop.

[1]: https://doc.rust-lang.org/core/ptr/fn.write_volatile.html

Test: m pvmfw # Then check disassembly for STRB
Change-Id: I6d404df4c0439b8e208d1ef533b60434d2b2bec8