Revert "Add mock settings pages to terminal app"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7d2867d6-5118-4b23-b8a0-152c0f0276a5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If6dca5cd21f5e6c10785341f3a6f09561548ed73
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com