Revert "Add mock settings pages to terminal app"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/de2e6024-2885-499d-89c4-a9428600cc70). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0f05d1160483cdd35e43b0c9c781351ba132869c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com