Revert "refactor build script"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5fa0232c-9e30-4f1f-b7d3-9d0e74348644). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1967c1afe4422d7c2ac51a4ef62a73b0b0e55f87
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com