Revert "refactor build script"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/04082ed9-9313-45e4-ae77-5a2442649901). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4fcb7de38d22dc991099021f5187622b73a87afb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com