virtmgr: migrate from vm_control to crosvm_control
`vm_control` is a crosvm internal API that is subject to arbitrary
refactoring. `crosvm_control` is the stable public API equivalent.

It is no longer possible to detect the `ENOTSUP` case when requesting
balloon stats. Contrary to what the comment said, that case only occurs
when the VM was configured without a virtio-balloon device, so we
preverse the old behavior by tracking whether we added the balloon.

If the balloon device is present, but hasn't been initialized by the
guest yet, the call will simply block until it has been initialized
(which is possibly undesirable and needs rethinking).

Test: m
Bug: 366271643
Change-Id: Ic58c3c9cf44cb06d517f8739d73f845d22832609