Revert "liblibfdt_bindgen: Only use libfdt as header_libs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c309347d-460d-4257-9458-8d41c5993357). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I763de4373b74dbce0da6f3786ea0df0eb275edee
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com