Revert "liblibfdt: Add no_std variant for libfdt_baremetal"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c309347d-460d-4257-9458-8d41c5993357). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7b26716e648338a7d791bd9556403c0275f635e0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com