Revert "Add default account "droid""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/22dfa43d-2c67-4051-adc5-0b39df1c96f6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I22f03624b8422d9fef82931028fcb691063e15ec
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com