Revert "VmTerminalApp: Merge with LinuxInstaller"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a33aebdc-d282-41c0-8259-dbe5ba4a9b20). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3d09b21d5d1d96b72dcdd3071370cce0b4625a28
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com