Add microdroid_kernel_16k prebuilt
These kernel prebuilt is conditionally added to the com.android.virt
APEX based on the value of the target_boots_16k soong config variable.

This patch uses a locally built 16k microdroid kernel. Once the CI build
is set up, we will switch to the prebuilt built on the build servers.

Bug: 317201718
Test: builds
Test: adb shell /apex/com.android.virt/bin/vm info
Change-Id: Ib815c3f7e67044e9d5ee6dde22f969505b765389