Revert "VmTerminalApp: Merge with LinuxInstaller"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2077989e-390d-4eb7-b3fa-b0597cfce35d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8ef20dd5ad9dfe71efb4ef644436292c7cc37582
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com