pvmfw: Map image footer after dynamic PT switch
Teach the MemoryTracker to map the optional image footer into the
dynamic PTs. As vmbase now knows internally where it is located, there
is no need to receive the "payload range" from the client anymore.
Therefore, refactor MemoryTracker::new() to stop taking that argument.

As a result, pvmfw now first switches to the dynamic PTs before
requesting the footer to be mapped, which prepares it (along with other
vmbase clients) for an upcoming change that will activate the dynamic
PTs before it is entered from vmbase.

As the MemoryTracker keeps track of whether the footer has been mapped,
get_appended_data_slice() can now safely return a Result<&mut> as it
will fail if called more than once.

Note that the memory tracker uses dirty-bit tracking (flushed when
switching back to the static PTs) for the image footer.

Bug: 377276983
Test: m {pvmfw,rialto,vmbase_example_{bios,kernel}}_bin
Test: atest rialto_test vmbase_example.integration_test
Change-Id: I24954b6905db0a0a499f7aa2dedb9a4e5b83d168