pvmfw: Switch to dynamic PT, then map image footer
As the region is now known to vmbase, extend the newly introduced memory
API with a function abstracting away the operation of locating/mapping
the region before mapping it. Do this separately from the static PT
replication, as most clients won't expect the footer.

Use this new function in pvmfw. As a result, we can now map the region
after having switched to the dynamic PT, enabling an upcoming change
that will enter pvmfw with the dynamic PT live.

For other clients, refactor switch_to_dynamic_page_tables() to remove
the need to inform it that no payload is expected (None).

Note that the memory tracker uses dirty-bit tracking (flushed when
switching back to the static PTs) for the image footer.

Bug: 377276983
Test: m {pvmfw,rialto,vmbase_example_{bios,kernel}}_bin
Test: atest rialto_test vmbase_example.integration_test
Change-Id: I24954b6905db0a0a499f7aa2dedb9a4e5b83d168