Revert "Add waitForBootCompleted to the Terminal's MainActivity"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/886ee62a-6e72-48fd-9168-082a85a85b67). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icec8c5d8119bc702c4d691d57248ea9ebc75f313
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com