Revert "Use /data/user/<user_id>/<package_name> directory"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/81d742b8-57a1-44bf-94a9-c7e8097150e2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I20b7650addc03994d7d318e7ead802b32d4952a0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com