Revert "Some UI change to the terminal app"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d4d09311-9987-486b-80fa-b9b884b0af14). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7a85aa008461f9456c2c45b23e07baa07eee0535
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com