Revert "Use /data/user/<user_id>/<package_name> directory"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8f2696b2-58fd-4c69-a6aa-f7cc7047780f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I25215882e93b705f35c64390eb2a96de08312841
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com