Revert "Don't put the terminal app on the devices with no AVF"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/19d7aae0-0b08-4023-ba3f-af80323f131d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I61bdf5041a5d7584805174b10488c7aa32b27108
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com