Revert "Use /data/user/<user_id>/<package_name> directory"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/617af094-3739-4214-9c56-4355bf4f1195). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idb8da79c3763009650b85252d4c5bc057141513c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com