Revert "Don't put the terminal app on the devices with no AVF"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/408b0e6e-7201-41e9-92ae-d67c07563ca0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifaa1d800cd0b0be6bd6137aa3ad309ab6adf15f1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com