Revert "Add VSR indexes for VM remote attestation VTS"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/938ae5b4-246b-4d53-8a7b-a628fa547c4f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib80bd316f3f887bb1ccdbf52576f133958fcace0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com