Revert "Propagate RELEASE_AVF_ENABLE_VM_TO_TEE_SERVICES_ALLOWLIS..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/23442cea-6e1c-43cf-879f-44fd9af2e737). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 379196597
Change-Id: I5edc0a9a8c8d69d955223b230389b358c989a347
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com