Revert "Implement permission check for VMs that request tee_services"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6d2fd6cd-c195-420d-924e-0171b7e54930). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 379208325
Bug: 379208127
Bug: 379208625
Change-Id: Id11d2f10214f409ed85a393ed7a0a68e765b387b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com