Revert "virtmngr: get secontext of the caller app"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ed14978d-1453-42c4-8eff-c15a32c7e2d0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 379208798
Change-Id: Ibd9f5ec9f1f51eba9fa82986c001875bf876b0d8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com