Revert "Implement permission check for VMs that request tee_services"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ed14978d-1453-42c4-8eff-c15a32c7e2d0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 379208798
Change-Id: I2141327952331fb7a7dccb24d1e91f441679bece
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com