Revert "Propagate RELEASE_AVF_ENABLE_VM_TO_TEE_SERVICES_ALLOWLIS..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6d7deaf5-9333-485e-a864-7e07a2eb8eea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 379159866
Change-Id: I7e0d51c7097bf10aa0921aa13ca2b0502b192206
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com