Revert "vm cli: add --tee_services flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6d7deaf5-9333-485e-a864-7e07a2eb8eea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 379159866
Change-Id: If9ac6561503d8c16165fa427b4de05de996b7601
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com