Revert "Remove unnecessary OWNERS file in libs/dice"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1d4e145a-0133-4dff-a2a4-22a2128cf568). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iff5ba6fd6523204bd858d540f808d90b8f613414
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com