Revert "Fix icons in dark mode"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/66cc9b59-3ecf-42d9-8c9b-21cfae803073). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id67af3fe11a8986c1af5bc583c1e53475b9bba5a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com