Revert "Use /data/user/<user_id>/<package_name> directory"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d6aae8f7-0655-4c55-9074-08a2f4eaab51). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7b0a51dc61a7f29ec2a7eec0464cb9374a58483c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com