pvmfw: Don't abort boot for missing /avf/untrusted
The node is only needed when booting a VM that is rollback-protected
through either deferred RBP (Secretkeeper) or fixed rollback index (RKP
VM) and would cause a "legacy" VM relying on the instance.img to fail
booting, if missing. This is not justifiable so treat the node missing
in the same way the property missing is handled i.e. by falling back to
instance.img.

Bug: 291213394
Bug: 377276983
Test: m pvmfw_bin
Change-Id: Iea4560e532e3d03722713ee01aa481be95f8f264