pvmfw: Don't abort boot for missing /avf/untrusted
The node is only needed when booting a VM that is rollback-protected
through either deferred RBP (Secretkeeper) or fixed rollback index (RKP
VM) and would cause a "legacy" VM relying on the instance.img to fail
booting, if missing. This is not justifiable so change the logic to
fallback to "legacy" RBP when the node is missing.

Take this opportunity to refactor FDT ops out into fdt.rs.

Bug: 377276983
Test: m pvmfw_bin
Change-Id: Iea4560e532e3d03722713ee01aa481be95f8f264