Revert "Define interface for reporting active ports in the guest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7e87f16c-5cef-44ca-b146-3089be72e3f6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If053bb284da3850d164f81643d5c134af8bb4f4e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com