pvmfw: Always use instance.img if !llpvm_changes
As build flags are intended to prevent the use of merged functionality
until flipped, remove the logic making pvmfw panic when encountering an
image using the newly-introduced capabilities and llpvm_changes wasn't
enabled at compile time. Instead, fall back to the legacy behavior of
relying (unconditionally) on the instance.img.

Refactor the instance.img path to make it (hopefully) simpler to follow.

Bug: 377276983
Test: m pvmfw_bin
Change-Id: I4db0655673a3186897ac465ab66b95cc32434b08