pvmfw: Fix RBP when !cfg!(llpvm_changes)
Build flags were introduced in aosp/2985312 to guard the new support for
Secretkeeper-backed deferred RBP in pvmfw. However, the resulting
behavior of pvmfw when the flag is NOT enabled also ended up being
altered, even though it should have remained unchanged.

In particular, the RKP VM was previously receiving a DICE chain derived
with an instance.img-backed salt. However, after the introduction of the
llpvm_changes flag, a pvmfw build where !cfg!(llpvm_changes) would panic
when encountering the RKP VM as it unwrapped the instance_hash.

Similarly, pvmfw became unconditionally aware of the new "capabilities"
and, as those _require_ the instance_hash, would crash when encountering
payloads containing them if the build flag wasn't set.

Therefore, only support SecretkeeperProtection and the very recent
TrustySecurityVm if cfg!(llpvm_changes) and otherwise fallback to the
legacy behavior for both the RKP VM (as originally intended in
aosp/2868574) and "regular" instance.img-backed VMs.

Clarify in perform_legacy_rollback_protection() that the instance_hash
is required and the only value used as salt when cfg!(llpvm_changes).

Bug: 291213394
Bug: 377276983
Test: m pvmfw_bin
Change-Id: I4db0655673a3186897ac465ab66b95cc32434b08