pvmfw: Read untrusted DT directly in early code
Stop relying on the output of the DT sanitization for configuring the
memory region and ranges for the kernel and ramdisk as those will be
validated by the MemoryTracker anyway. This will enable moving the
sanitization to main(), to perform it with more function-level inputs.

Note: No functional change intended.

Bug: 377276983
Test: m pvmfw_bin
Change-Id: Id7f24fc6952dcc508aee04da04db9ed15b921036