pvmfw: Move DT sanitization to main()
Move it so that it can re-use other inputs e.g. VBMeta capabilities.
This should not compromise the integrity of pvmfw as it still happens
before the DT is used to initialize the PCI bus and swiotlb region.

Move the shared memory setup to main(), so that it only happens when
actually needed i.e. when the instance.img is used, requiring a shared
region for the VirtIO transport.

Bug: 339782511
Bug: 377276983
Test: m pvmfw_bin
Change-Id: I6de6e4607d110c2f8fffd9a3d9527716dca92565