Revert "Use /data/user/<user_id>/<package_name> directory"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/63859b49-a0c5-428d-9361-3e2441a5a874). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I07fca0c7283f2b09e83a8fa5464d704d3a36f0c0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com