Revert "Debian service should check its caller is the vm"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/efd5113b-e620-4996-b4ac-39907cb23cea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I65aee3b222148524e9ed7a59e4f03710858c6f95
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com