Revert "Add a TODO to remove python3 -u"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/abf86674-9d00-4f6f-831d-5db6b0a97576). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380450536
Change-Id: I01e3d98fda7fdb2953a84bb9aed48d0c8c01e76e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com