Revert "Implement permission check for VMs that request tee_services"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d70ea2a3-d1ba-454a-981f-e47254a4ffb5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibc374eaa66e53dfbcf6f14a408e338267fa7bd7f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com