Revert "Propagate RELEASE_AVF_ENABLE_VM_TO_TEE_SERVICES_ALLOWLIS..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d70ea2a3-d1ba-454a-981f-e47254a4ffb5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I004ebdebf61c31e818a24af30a95be14feb9cb9e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com