Revert "Remove unnecessary OWNERS file in libs/dice"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1ee8ed5f-d038-4c14-b0b6-a89b23cf1d68). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I905f3669f60c27bf017c0ba41dcea4e1e09b7fb1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com