Revert "Implement backup in terminal"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0fcb5137-ac86-41a6-853f-6504efada5ec). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8a24dbfe6b7f0a5e781250bc9e4b4fa982ea050f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com