Revert "Add early_vms.xml to Virtualization package to allow reu..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a99fb312-c8e2-4ec3-816a-c78daed5d694). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic99c8224ef128b72559310c95b4f2b52938e936f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com