vmbase: Fix undetected EH stack overflows
If pvmfw receives a payload, it maps the region potentially containing
it (from ALIGN(bin_end, 4096) to ORIGIN(image) + LENGTH(image)) with
read-write permissions, meaning that the EH stack could overflow into it
and go undetected.

Address this by reserving a page of the writable_data region as guard
page instead and move the EH stack appropriately. Alternatively, we
could have "reserved" the last page of the image but this risks causing
unexpected issues when we say that the region available to the loader is
2 MiB but then unexpectedly crop it by one page.

Similar to the .stack, give the guard page its own section for clarity.

Bug: 381440625
Test: m {pvmfw,rialto,vmbase_example_{bios,kernel}}_bin
Change-Id: I17c32d5f8b5fae4fb7af325c05de3c08047922c1