Revert "forwarder_host terminates when terminating VM"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dc7c864f-6ad1-415c-ab6a-780e121fb951). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7e74b2d4f043b2a586e50bc1be0ea92ac01eb25d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com