Revert "Fix Ferrochrome bootloop"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/58405f02-703f-4516-9c0e-c8a852fbbcb2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I33846ca801dd09c4a3511d132c016bb174c70f48
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com