Revert "Move launchErrorActivity into ErrorActivity"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/58405f02-703f-4516-9c0e-c8a852fbbcb2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I338ef2efae66e158e124233c18eb88c1b4e64dec
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com