pvmfw: Support com.android.virt.name property
Teach pvmfw to recognize a new AVB property that will be used to gate
features specific to particular VMs, such as the way we currently
support the RKP VM with special rollback protection and its DICE chain
marker. For now, only add the infrastructure and unit-tests.

Limit the maximum size of the name to 16 bytes, so that it would fit in
low-level structures such as the newly-introduced config data entry for
"reserved memory" (aosp/3466214).

Note: No functional change intended.

Bug: 378673494
Bug: 377276983
Test: m pvmfw_bin
Test: atest libpvmfw_avb.integration_test libpvmfw.dice.test
Change-Id: Ic1fd923df361e19b2d0cd323aa6a0ca866a281a6