pvmfw: Use com.android.virt.name for RKP VM
Rely on this newly-introduced property to identify the RKP VM instead of
the soon-to-be-deprecated Capability::RemoteAttest, in order to select
the fixed-index RBP solution and control whether the RKP VM DICE chain
marker should be set.

Note that defining this property also results in pvmfw setting the DICE
component_name to it, which will be recognized by the RKP server. Thanks
to the specialized RBP solution, pvmfw would reject any other VM
attempting to re-use this name.

Bug: 378673494
Bug: 390527025
Bug: 391620545
Test: m pvmfw_bin
Test: atest libpvmfw_avb.integration_test libpvmfw.dice.test
Change-Id: I1a839501fb822006ed5f38c354ebe0d55948774f