Revert "add modifier keys in terminal app"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/81bf14ec-4f2a-4a4d-8b1e-62142d472dc2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If848fa79370ba497dcdb948c2785169003d086f8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com