Only interact with secretkeeper in updatable VMs
Change I893604ce5fc25d3e1c7261afa0de915902154be6 introduced a call to
secretkeeper to fetch its public key from v2 of the HAL (if available)
as opposed to from the device tree.

This is a problem for VMs that run on early boot, when secretkeeper is
not yet available.

Only attempt to access secretkeeper if an updatable VM is requested,
which should only happen in VMs that run when secretkeeper *is*
available.

Bug: 382376014
Test: TH
Change-Id: I1c27ca5981500dc4ccaa0677ae61d0434b2d7653