Revert "microdroid: Add kernel image variant that supports UEFI"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/34f70eea-4bd2-445d-a5ae-b710bed78e8f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic69c3e07ebecadbbb7551353cd68b52fac516789
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com