libavf: remove AVirtualMachineRawConfig_setBalloon
This API isn't used yet, so it is safe to delete.

virtio-balloon is now disabled by default and initial users of this API
don't need it, so remove it to keep the API small. One reason to avoid
exposing balloon support at this point is that a single bool doesn't
capture all the potential balloon features, e.g. a client may or may not
want free page reporting enabled.

Test: m
Bug: 369588412
Change-Id: I10872f29a34b6a20777e985fa97d21e7bd26cc6f