Revert "microdroid: Add kernel image variant that supports UEFI"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/29939eb6-d4a9-4716-9398-b2049339b209). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I95bb7290b053d582df55d89a1718234eaaea4f65
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com