build/debian: Verify checksums and signature for kernel sources
In this CL, we add some basic validation for the kernel sources, before introducing our own patches, and finally building the kernel.

Specifically, we verify the source package checksums and that the source package contains a GPG signature that can be be verified using `debian-keyring`. The build fails if either of these conditions are unmet.

Bug: 380176863
Test: ./build_in_container.sh -k
Test: ./build_in_container.sh -a aarch64 -k

Change-Id: I235ff507b03e368183fa44737db6ab34ec0a7ddc