Revert "Introduce JavascriptAddon"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f4b96252-864a-4a5e-afd3-dbbb5fcc9a01). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I73cc51782ffa995843520e9e4e989c50169df126
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com