Revert "pvmfw: Support com.android.virt.page_size"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9fc1d9aa-400c-437c-836e-5bc7a12afd81). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2c0ac3abbf6d43f84a69fe99fbef18e7ecb1c11e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com