Revert "Add null check for PortNotifier"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c6b5c466-e817-4d49-b394-30ef804bccf7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie26a9b59fead39698852499ef0f64b02cdffaa35
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com