Revert "pvmfw: Move RBP into crate::rollback"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9e648e47-a34f-4a87-867f-7b1250fdd695). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibdc183d487953d636114ee5bfe0299b9710ab379
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com