Add option to start trusty from raw binary image
This change adds option to start trusty from raw binary image. This is
required for trusty image to be able to be verified by pvmfw. pvmfw
relies on AVB footer for verification, which currently cannot be used
with an elf image. Futhermove Trusty currently does not support other
kernel image format to be verified by the crosvm.

Bug: 375584439
Test: /system_ext/bin/trusty_security_vm_launcher --raw-binary
Change-Id: Id72efcb95ed992e4e0270d4e21de8e2f5e74f9be