Revert "build/debian: Verify checksums and signature for kernel ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fd918c57-72ec-4ac5-ab6a-5d9915488f5d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie58461ccd8fca6d693d719b2c1381900b06e7075
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com