Revert "VmTerminalApp: Limit disk resize up to free disk size"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4d9fe0ca-83ad-49de-9a24-d014bf958151). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I311b4f7026ef1e2767e045da7291e67b67c7007d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com