Revert "add option for custom vCPU count"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/34807d68-8dbd-410d-8dd3-a30baaafebcb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8993870cc2cb468e3d3bad2055f1f2caf9381942
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com