Revert "Convert ErrorActivity+a to kotlin"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/50874627-d155-4d9a-b69b-f12f3c1c702d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 385686293
Change-Id: I48b90f3220723e7fcb35a6be6a137b602d656acf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com