Revert "Convert ErrorActivity+a to kotlin"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6ed5aff8-7052-4085-8f1f-672fdae6f91d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 385692932
Bug: 385692932
Change-Id: Ie57052bc44dbd74048a8bba46f7f1e2c851e7625
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com