Revert "Convert ErrorActivity+a to kotlin"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7b8bf1bf-f976-4ea5-9e96-ccd8a64ac119). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 385692788
Change-Id: I01c787f9ec164aff433a82fecc21a31d3ae99c3e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com