Revert "Convert ErrorActivity+a to kotlin"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a1169d72-42df-4c5c-8da9-fccef4e15e4d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 385703347
Change-Id: I3789ea818ff15e9c4f364fb1c573676130c4adc8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com