Revert "Convert ErrorActivity+a to kotlin"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/69313bb5-b5f4-4d1a-a32f-3b8b73398fc0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 385692217
Bug: 385692217
Change-Id: I1839af48aa2fecb1c1b8601901a43c8fc9794def
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com