Revert "Convert ErrorActivity+a to kotlin"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c027ca20-16cc-4a77-9bb3-7752de3c475c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 385692275
Bug: 385692275
Change-Id: I014fe54d878e1fb8bc71eb6b7463b995e85518ad
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com