Revert "Print debug messages from LLNDK to logcat"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/83fe8037-3287-41a1-8504-bbc2fc3d990b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib57301bde33e6e90eb096007bfe7acdba297bf35
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com