Revert "Check |ro.vendor.api_level| for vendor API level on GRF/..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4adf0173-ff42-4dbb-b7a3-4d45803beb0a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2d8e9b6b5f3b6566d1f9227a8585f931054111fb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com