Revert "Show stack trace in the error activity"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4adf0173-ff42-4dbb-b7a3-4d45803beb0a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia267b5ee5eba3ed218a2681e5c0693ac2f1a066f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com