Revert "libavf: Clarify what the “instanceId” is."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/464c977e-22bb-443e-9595-364a8ead077c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I330e0d2ae0702c354d0241f933000bd586f4ca8c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com