Revert "Convert TerminalView to kotlin"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8c546ad1-c39f-4bad-91fb-f82d3209ad0d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib0e5bb8f5af28b1dfe2d54249dd1eb2b3bc56065
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com