Revert "libavf: Clarify what the “instanceId” is."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d52d5b8e-bd8d-4e03-9aea-e14612677403). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9ed1ef5d1c483e83bb0a76e60dccb596bff549f0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com