Revert "Convert Installer* and BaseActivity to kotlin"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f69337eb-b62f-4448-8c21-f0f4fc2df9a0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 387214951
Change-Id: I6ee78cbeaa23d7b457600cf9d85cf7663e8698a8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com