Revert "Refine schema of ReportVmActivePortsRequest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4c30226d-4403-4a9d-8f71-82dabc3ddbdc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0ca6b205826337a0f65a2f6f10ca5833b9c4633c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com