Revert "use android.R.id as the original code does in InstallerA..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/08318a02-2c5a-466b-b980-5f6cd5c5ad4d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id0ea4999a6beb9e49a2a7d60775e32f57ac46b0f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com