Revert "Refactor some field names in kotlin classes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/10dcf0d8-77de-4369-b8cc-d027cd0c7321). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia0355d77242443410132ff68c7d89c627f75f268
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com