Revert "Refactor some field names in kotlin classes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/226fc097-aa9e-4d1d-9187-736f5fd89786). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I46fbac3243be3c538b411a4a2938ef2b28b1343f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com