Revert "Refactor some field names in kotlin classes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e6201477-32f7-4e8b-81c0-d33e4df1efc0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I27cc15dc05b188a8ca8474d873d89584bd300488
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com