Revert "Refactor some field names in kotlin classes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dcfbc2c6-18bd-4863-838f-4a90532f0211). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1d28108f18165bca7c73a144ba674ffa381c5208
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com