Revert "Refactor some field names in kotlin classes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1f201223-f4e5-4b62-b526-6a35f76883b7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4a990340af6578f6ca0687ce66c8ea5196f5a922
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com