Revert "Refactor some field names in kotlin classes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3f25e931-f7c9-40e4-8566-fb671ed8e4a1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I87becc1629f83b22108c10983fb97a1f4471a97c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com