virtmgr: Don't detect hangup when using GDB
Currently, virtmgr times out and terminates the VM if something has
gone wrong during the boot of the guest kernel, preventing it from
properly reaching a point where it can communicate with the host.
However, if GDB is in use, it is likely that the kernel is being stepped
through, held on a breakpoint, ... Timing out in such conditions is not
only wrong but very unhelpful to the debugging session taking place.

Therefore, don't "detect hangup" when a GDB port has been passed.

Test: TH
Bug: 385659491
Change-Id: Ie11597070fbfc8c72b357bc9746ad5d5789521eb