Revert "microdroid: Add kernel image variant that supports UEFI"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7799e56a-0905-4869-8900-0ce077db238b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5e0b56c4438b39694edd574057b9b3c221479489
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com