Revert "dice: add trusty build rules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a9499393-3e7d-4654-b2e2-f6847b89d162). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 388797107
Change-Id: I69291ed3b87c0f19c4e18352ae1fd8ea1725e719
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com