Revert "dice: add trusty build rules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/59845ac0-67b0-4077-a61f-2aa776bd7a0e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 388812616
Change-Id: I26f62e210b176d3d9524b06ba0f8b87936e76f24
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com