Revert "dice: add trusty build rules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fec45dc4-f16c-4c0d-953f-cba751f0d17e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 388813183
Change-Id: I8bcf90cab3993106f9bb1c1a988e1c891069e48d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com