Revert "DebianServiceImpl to Kotlin"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2c4223db-a984-4fe5-bffc-32d0665ff5e0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I62a4247f65f2517101ef45361c4d271165826393
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com